752 Junit tests with updated schema

+++ tools/repository/server/src/main/java/eu/europa/ec/leos/repository/repositories/DocumentCategoriesVRepository.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 European Commission
+ * Copyright 2023 European Commission
  *
  * Licensed under the EUPL, Version 1.2 or – as soon they will be approved by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
@@ -16,5 +16,7 @@ package eu.europa.ec.leos.repository.repositories;
 import eu.europa.ec.leos.repository.entities.DocumentCategoriesV;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentCategoriesVRepository extends JpaRepository<DocumentCategoriesV, Long> {
+import java.math.BigDecimal;
+
+public interface DocumentCategoriesVRepository extends JpaRepository<DocumentCategoriesV, BigDecimal> {
 }
